add SelectedWalletMetaContext
