--- legacy/components/WalletSelection/styles/WalletSelectionScreen.style.js
@@ -1,34 +1,0 @@
-// @flow
-
-import {StyleSheet} from 'react-native'
-
-import {COLORS} from '../../../styles/config'
-
-export default StyleSheet.create({
-  safeAreaView: {
-    flex: 1,
-    backgroundColor: COLORS.BACKGROUND_BLUE,
-  },
-  container: {
-    flex: 1,
-  },
-  title: {
-    textAlign: 'center',
-    fontSize: 24,
-    color: '#fff',
-    paddingVertical: 16,
-  },
-  wallets: {
-    margin: 16,
-    flex: 1,
-  },
-  topButton: {
-    marginHorizontal: 16,
-    marginTop: 16,
-    marginBottom: 10,
-  },
-  button: {
-    marginHorizontal: 16,
-    marginBottom: 10,
-  },
-})
