+++ src/SelectedWallet/WalletSelection/WalletSelectionScreen.stories.js
@@ -1,9 +1,7 @@
-// @flow
-
 import {storiesOf} from '@storybook/react-native'
 import React from 'react'
 
-import WalletSelectionScreen from './WalletSelectionScreen'
+import {WalletSelectionScreen} from './WalletSelectionScreen'
 
 storiesOf('WalletSelectionScreen', module).add('Default', ({navigation}) => (
   <WalletSelectionScreen navigation={navigation} />
