+++ src/SelectedWallet/WalletSelection/index.ts
@@ -1,0 +1,1 @@
+export * from './WalletSelectionScreen'
